+++ zakozhurnikova.kristina/P1/main.cpp
@@ -1,38 +1,7 @@
 #include <iostream>
-#include <limits>
 #include <stdexcept>
+#include "sequence_counter.hpp"
 
-struct SequenceCounter
-{
-  SequenceCounter():
-   counts(0)
-  {}
-
-  void count(int number)
-  {
-    int max = std::numeric_limits< int >::min();
-    size_t max_size = std::numeric_limits< size_t >::max();
-    if (counts == max_size)
-    {
-      throw std::logic_error("sequence is too long");
-    }
-    else if (number > max)
-    {
-      counts = 0;
-      max = number;
-    }
-    else
-    {
-      ++counts;
-    }
-  }
-  size_t get_result() const
-  {
-    return counts;
-  }
-private:
-  size_t counts;
-};
 int main ()
 {
   int number = 0;
